OneDrive/calenderWeb/calenderwebbackend/pyproject.toml
[tool.poetry]
name = "calenderwebbackend"
version = "0.1.0"
description = ""
authors = ["alsktv <alsktv@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
pillow = "^10.3.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.235.0"}
strawberry-graphql-django = "^0.44.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
